+++ yellowstone-grpc-geyser/src/block_reconstruction.rs
@@ -186,6 +186,7 @@ pub struct BlockMachineStorage {
     ready_queue: VecDeque<(SlotCommitmentStatusUpdate, Arc<FrozenBlock>)>,
     state: BlocksStateMachine,
     min_slot: Option<u64>,
+    num_buffered_finalized_slot: usize,
 }
 
 pub struct ReplayIter<'storage> {
@@ -233,6 +234,21 @@ impl<'storage> Iterator for ReplayIter<'storage> {
     }
 }
 
+///
+/// As blocks are frozen, they are added to the `replayed_slot` map.
+///
+/// This map has a maximum capacity, and when that capacity is exceeded, the oldest slots are pruned.
+///
+/// The [`MINIMUM_FINALIZED_SLOT_TO_BUFFER`] constant makes sure we have enough finalized slot in our state to avoid pruning slots
+/// that are still in the process of being finalized, which could lead to missing finalized/confirmed messages for those slots.
+///
+/// See [`BlockMachineStorage::gc`] for more details.
+///
+///
+/// We used to do this in prior to version v13.2.0 of the plugin (before the reconstruction refactor)
+///
+pub const MINIMUM_FINALIZED_SLOT_TO_BUFFER: usize = 10;
+
 impl BlockMachineStorage {
     pub fn new(replayed_capacity: usize) -> Self {
         Self {
@@ -244,6 +260,7 @@ impl BlockMachineStorage {
             ready_queue: VecDeque::new(),
             state: BlocksStateMachine::default(),
             min_slot: None,
+            num_buffered_finalized_slot: 0,
         }
     }
 
@@ -251,7 +268,12 @@ impl BlockMachineStorage {
         self.processing_slots.remove(&slot);
         self.pending_blockmeta.remove(&slot);
         self.replayed_slot.remove(&slot);
-        self.slot_commitment_progression_map.remove(&slot);
+        if let Some(progression) = self.slot_commitment_progression_map.remove(&slot) {
+            if progression.max_commitment == CommitmentLevel::Finalized {
+                self.num_buffered_finalized_slot =
+                    self.num_buffered_finalized_slot.saturating_sub(1);
+            }
+        }
     }
 
     fn slot_reset(&mut self, slot: u64) {
@@ -356,7 +378,9 @@ impl BlockMachineStorage {
     }
 
     fn gc(&mut self) {
-        while self.replayed_slot.len() > self.replayed_capacity {
+        while self.replayed_slot.len() > self.replayed_capacity
+            && self.num_buffered_finalized_slot > MINIMUM_FINALIZED_SLOT_TO_BUFFER
+        {
             if let Some((&oldest_slot, _)) = self.replayed_slot.iter().next() {
                 self.prune_slot(oldest_slot);
             }
@@ -383,9 +407,24 @@ impl BlockMachineStorage {
                 self.slot_commitment_progression_map
                     .entry(slot)
                     .and_modify(|progression| {
-                        progression
+                        // Make sure its not there already
+                        if !progression
                             .commitment
-                            .push(slot_commitment_status_update.clone());
+                            .iter()
+                            .any(|s| s.commitment == slot_commitment_status_update.commitment)
+                        {
+                            progression
+                                .commitment
+                                .push(slot_commitment_status_update.clone());
+
+                            if matches!(
+                                slot_commitment_status_update.commitment,
+                                CommitmentLevel::Finalized
+                            ) {
+                                self.num_buffered_finalized_slot += 1;
+                            }
+                        }
+
                         if cmp_commitment_level(
                             slot_commitment_status_update.commitment,
                             progression.max_commitment,
@@ -394,20 +433,30 @@ impl BlockMachineStorage {
                             progression.max_commitment = slot_commitment_status_update.commitment;
                         }
                     })
-                    .or_insert_with(|| SlotProgression {
-                        commitment: vec![slot_commitment_status_update.clone()],
-                        max_commitment: slot_commitment_status_update.commitment,
+                    .or_insert_with(|| {
+                        if matches!(
+                            slot_commitment_status_update.commitment,
+                            CommitmentLevel::Finalized
+                        ) {
+                            self.num_buffered_finalized_slot += 1;
+                        }
+
+                        SlotProgression {
+                            commitment: vec![slot_commitment_status_update.clone()],
+                            max_commitment: slot_commitment_status_update.commitment,
+                        }
                     });
 
                 let commitment_level = slot_commitment_status_update.commitment;
-                if matches!(commitment_level, CommitmentLevel::Finalized) {
-                    // Only gc on finalized, that should be enough.
-                    self.gc();
-                }
 
                 if let Some(frozen_block) = self.replayed_slot.get(&slot) {
                     self.ready_queue
                         .push_back((slot_commitment_status_update, Arc::clone(frozen_block)));
+                }
+
+                if matches!(commitment_level, CommitmentLevel::Finalized) {
+                    // Only gc on finalized, that should be enough.
+                    self.gc();
                 }
             }
             BlockStateMachineOutput::ForksDetected(fork_detected) => {
@@ -931,5 +980,23 @@ mod tests {
             /* Last message is explicitly sent as BlockMeta, its no longer implicit inside of original_messages */
             assert_eq!(entry_count, 2);
         }
+    }
+
+    #[test]
+    fn block_machine_should_yield_all_block_and_commitment_when_replayed_capacity_set_to_zero() {
+        let mut storage = BlockMachineStorage::new(0);
+        drive_slot_to_processed(&mut storage, 1, None);
+        storage.add(make_slot_msg(1, None, SlotStatus::Confirmed));
+        storage.add(make_slot_msg(1, None, SlotStatus::Finalized));
+
+        // Drain the ready queue
+        let (actual, _block) = storage.pop_ready_block().unwrap();
+        assert!(actual.commitment == CommitmentLevel::Processed);
+
+        let (actual, _block) = storage.pop_ready_block().unwrap();
+        assert!(actual.commitment == CommitmentLevel::Confirmed);
+
+        let (actual, _block) = storage.pop_ready_block().unwrap();
+        assert!(actual.commitment == CommitmentLevel::Finalized);
     }
 }